chore(ci): Upgrade GitHub actions to v4
This commit updates the `actions/checkout` and `actions/upload-artifact` GitHub Actions from v3 to v4 in the OpenWrt build workflow.

The upgrade is necessary to stay on supported versions, as v3 of these actions relied on Node.js 16, which is now deprecated. Version 4 utilizes the Node.js 20 runtime, ensuring the workflow remains secure, efficient, and compatible with the latest GitHub Actions environment.